requirements.txt
requests==2.26.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
cssselect==1.1.0
idna==3.3
lxml==4.6.3
readability-lxml==0.8.1
urllib3==1.26.7
beautifulsoup4==4.10.0
soupsieve==2.2.1
markdownify==0.9.4
six==1.16.0